feat(web): real-time dashboard, job log viewer, component split
- useDashboard hook: EventSource on /events with debounced re-fetches,
  automatic fallback to 5s polling when the stream is unavailable, and
  background retries to return to live updates. Header shows a
  Live/Polling/Connecting status badge.
- Job log viewer: per-job Logs action opens a dialog tailing
  /jobs/{id}/log/content, auto-refreshing while open, with ANSI stripping
  and follow-on-output scrolling.
- Split the 38KB App.tsx into components/dashboard/* (header, overview,
  jobs/gpus/reservations/stats views, log dialog, shared widgets),
  lib/format.ts, lib/jobs.ts and hooks/useDashboard.ts.
- Proxy /events in the vite dev server.
