+++ web/src/components/dashboard/StatePanels.tsx
@@ -1,0 +1,55 @@
+import { ShieldAlert } from "lucide-react"
+
+import {
+  Card,
+  CardContent,
+  CardDescription,
+  CardHeader,
+  CardTitle,
+} from "@/components/ui/card"
+import { Skeleton } from "@/components/ui/skeleton"
+import { TableCell, TableRow } from "@/components/ui/table"
+
+export function EmptyRow({ columns, label }: { columns: number; label: string }) {
+  return (
+    <TableRow>
+      <TableCell colSpan={columns} className="h-24 text-center text-muted-foreground">
+        {label}
+      </TableCell>
+    </TableRow>
+  )
+}
+
+export function ErrorState({ message }: { message: string }) {
+  return (
+    <Card className="rounded-lg border-rose-200 bg-rose-50 text-rose-950 dark:border-rose-950 dark:bg-rose-950/30 dark:text-rose-100">
+      <CardHeader>
+        <CardTitle className="flex items-center gap-2">
+          <ShieldAlert className="size-5" />
+          API unavailable
+        </CardTitle>
+        <CardDescription className="text-rose-800 dark:text-rose-200">
+          {message}
+        </CardDescription>
+      </CardHeader>
+    </Card>
+  )
+}
+
+export function LoadingState() {
+  return (
+    <div className="grid gap-4 md:grid-cols-2 xl:grid-cols-4">
+      {Array.from({ length: 8 }).map((_, index) => (
+        <Card key={index} className="rounded-lg">
+          <CardHeader>
+            <Skeleton className="h-4 w-24" />
+            <Skeleton className="h-8 w-20" />
+          </CardHeader>
+          <CardContent>
+            <Skeleton className="h-4 w-32" />
+          </CardContent>
+        </Card>
+      ))}
+    </div>
+  )
+}
